move decompose_clifford  to synethsis
